crewai-cookbook/requirements.txt
crewai
honeyhive
python-dotenv
openai 